backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
requests==2.32.0
python-dotenv==1.0.1
httpx==0.27.0
openai==1.50.0
